Normalise a geometry to fit inside a unit square

This is a small helper function that will take an sf geometry and fit it inside the unit square (a square centered on 0 and ranging from -1 to 1 in both dimensions). The function will retain the aspect ratio of the geometry and simply scale it down until it fits.

st_normalize(st)

Arguments

st
An sf geometry such as sf, sfc, or sfg

Value

An object of the same type as st

Examples

library(sf)

#> Linking to GEOS 3.6.2, GDAL 2.2.3, PROJ 4.9.3

#> 
#> Attaching package: ‘sf’

#> The following object is masked from ‘package:transformr’:
#> 
#>     st_normalize

nc <- st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
st_bbox(nc)

#>      xmin      ymin      xmax      ymax 
#> -84.32385  33.88199 -75.45698  36.58965 

nc_norm <- st_normalize(nc)
st_bbox(nc_norm)

#> xmin ymin xmax ymax 
#>    0    0    1    1