Transition between data.frames containing sfc columns

This function is equivalent to tweenr::tween_state() except that it understands sf::sfc columns, as defined by the sf package. An sfc column is a column containing simple features and can this hold both points, lines polygons and more. tween_sf currently has support for (multi)point, (multi)linestring, and (multi)polygon types and requires that the transition is between compatible types (points-to-points, linestring-to-linestring, polygon-to-polygon). For (multi)linestring and (multi)polygon, the behavior is similar to tween_path() and tween_polygon() respectively, with each feature beeing run through the respective function with match = FALSE. For (multi)points it behaves more or less like tweenr::tween_state() except additional points are added as needed to make the to stages contain the same number of points. Points are added on top of existing points so it appears as if the points are divided into more.

tween_sf(.data, to, ease, nframes, id = NULL, enter = NULL, exit = NULL)

Arguments

.data

A data.frame to start from. If .data is the result of a prior tween, only the last frame will be used for the tween. The new tween will then be added to the prior tween

to

A data.frame to end at. It must contain the same columns as .data (exluding .frame)

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

id

The column to match observations on. If NULL observations will be matched by position. See the Match, Enter, and Exit section for more information.

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

Value

A data.frame containing intermediary states

Examples

library(magrittr)
star_hole <- poly_star_hole(st = TRUE)
circles <- poly_circles(st = TRUE)
spiral <- path_spiral(st = TRUE)
waves <- path_waves(st = TRUE)
random <- point_random(st = TRUE)
grid <- point_grid(st = TRUE)
df1 <- data.frame(geo = sf::st_sfc(star_hole, spiral, random))
df2 <- data.frame(geo = sf::st_sfc(circles, waves, grid))

tween_sf(df1, df2, 'linear', 30)
#>                          geometry .id     .phase .frame
#> 1  POLYGON ((0 1, 0.2938926 0....   1        raw      1
#> 2  LINESTRING (0 0, 0.00050334...   2        raw      1
#> 3  MULTIPOINT ((-0.2228926 -0....   3        raw      1
#> 4  MULTIPOLYGON (((0.4637158 0...   1 transition      2
#> 5  MULTILINESTRING ((-0.034482...   2 transition      2
#> 6  MULTIPOINT ((-0.2496894 -0....   3 transition      2
#> 7  MULTIPOLYGON (((0.4539942 0...   1 transition      3
#> 8  MULTILINESTRING ((-0.068965...   2 transition      3
#> 9  MULTIPOINT ((-0.2764862 -0....   3 transition      3
#> 10 MULTIPOLYGON (((0.4442726 0...   1 transition      4
#> 11 MULTILINESTRING ((-0.103448...   2 transition      4
#> 12 MULTIPOINT ((-0.303283 -0.2...   3 transition      4
#> 13 MULTIPOLYGON (((0.434551 0....   1 transition      5
#> 14 MULTILINESTRING ((-0.137931...   2 transition      5
#> 15 MULTIPOINT ((-0.3300798 -0....   3 transition      5
#> 16 MULTIPOLYGON (((0.4248294 0...   1 transition      6
#> 17 MULTILINESTRING ((-0.172413...   2 transition      6
#> 18 MULTIPOINT ((-0.3568766 -0....   3 transition      6
#> 19 MULTIPOLYGON (((0.4151078 0...   1 transition      7
#> 20 MULTILINESTRING ((-0.206896...   2 transition      7
#> 21 MULTIPOINT ((-0.3836734 -0....   3 transition      7
#> 22 MULTIPOLYGON (((0.4053862 0...   1 transition      8
#> 23 MULTILINESTRING ((-0.241379...   2 transition      8
#> 24 MULTIPOINT ((-0.4104702 -0....   3 transition      8
#> 25 MULTIPOLYGON (((0.3956646 0...   1 transition      9
#> 26 MULTILINESTRING ((-0.275862...   2 transition      9
#> 27 MULTIPOINT ((-0.4372671 -0....   3 transition      9
#> 28 MULTIPOLYGON (((0.385943 0....   1 transition     10
#> 29 MULTILINESTRING ((-0.310344...   2 transition     10
#> 30 MULTIPOINT ((-0.4640639 -0....   3 transition     10
#> 31 MULTIPOLYGON (((0.3762214 0...   1 transition     11
#> 32 MULTILINESTRING ((-0.344827...   2 transition     11
#> 33 MULTIPOINT ((-0.4908607 -0....   3 transition     11
#> 34 MULTIPOLYGON (((0.3664998 0...   1 transition     12
#> 35 MULTILINESTRING ((-0.379310...   2 transition     12
#> 36 MULTIPOINT ((-0.5176575 -0....   3 transition     12
#> 37 MULTIPOLYGON (((0.3567782 0...   1 transition     13
#> 38 MULTILINESTRING ((-0.413793...   2 transition     13
#> 39 MULTIPOINT ((-0.5444543 -0....   3 transition     13
#> 40 MULTIPOLYGON (((0.3470566 0...   1 transition     14
#> 41 MULTILINESTRING ((-0.448275...   2 transition     14
#> 42 MULTIPOINT ((-0.5712511 -0....   3 transition     14
#> 43 MULTIPOLYGON (((0.337335 0....   1 transition     15
#> 44 MULTILINESTRING ((-0.482758...   2 transition     15
#> 45 MULTIPOINT ((-0.5980479 -0....   3 transition     15
#> 46 MULTIPOLYGON (((0.3276134 0...   1 transition     16
#> 47 MULTILINESTRING ((-0.517241...   2 transition     16
#> 48 MULTIPOINT ((-0.6248447 -0....   3 transition     16
#> 49 MULTIPOLYGON (((0.3178918 0...   1 transition     17
#> 50 MULTILINESTRING ((-0.551724...   2 transition     17
#> 51 MULTIPOINT ((-0.6516415 -0....   3 transition     17
#> 52 MULTIPOLYGON (((0.3081703 0...   1 transition     18
#> 53 MULTILINESTRING ((-0.586206...   2 transition     18
#> 54 MULTIPOINT ((-0.6784383 -0....   3 transition     18
#> 55 MULTIPOLYGON (((0.2984487 0...   1 transition     19
#> 56 MULTILINESTRING ((-0.620689...   2 transition     19
#> 57 MULTIPOINT ((-0.7052351 -0....   3 transition     19
#> 58 MULTIPOLYGON (((0.2887271 0...   1 transition     20
#> 59 MULTILINESTRING ((-0.655172...   2 transition     20
#> 60 MULTIPOINT ((-0.7320319 -0....   3 transition     20
#> 61 MULTIPOLYGON (((0.2790055 0...   1 transition     21
#> 62 MULTILINESTRING ((-0.689655...   2 transition     21
#> 63 MULTIPOINT ((-0.7588287 -0....   3 transition     21
#> 64 MULTIPOLYGON (((0.2692839 0...   1 transition     22
#> 65 MULTILINESTRING ((-0.724137...   2 transition     22
#> 66 MULTIPOINT ((-0.7856255 -0....   3 transition     22
#> 67 MULTIPOLYGON (((0.2595623 0...   1 transition     23
#> 68 MULTILINESTRING ((-0.758620...   2 transition     23
#> 69 MULTIPOINT ((-0.8124224 -0....   3 transition     23
#> 70 MULTIPOLYGON (((0.2498407 0...   1 transition     24
#> 71 MULTILINESTRING ((-0.793103...   2 transition     24
#> 72 MULTIPOINT ((-0.8392192 -0....   3 transition     24
#> 73 MULTIPOLYGON (((0.2401191 0...   1 transition     25
#> 74 MULTILINESTRING ((-0.827586...   2 transition     25
#> 75 MULTIPOINT ((-0.866016 -0.0...   3 transition     25
#> 76 MULTIPOLYGON (((0.2303975 0...   1 transition     26
#> 77 MULTILINESTRING ((-0.862069...   2 transition     26
#> 78 MULTIPOINT ((-0.8928128 -0....   3 transition     26
#> 79 MULTIPOLYGON (((0.2206759 0...   1 transition     27
#> 80 MULTILINESTRING ((-0.896551...   2 transition     27
#> 81 MULTIPOINT ((-0.9196096 -0....   3 transition     27
#> 82 MULTIPOLYGON (((0.2109543 0...   1 transition     28
#> 83 MULTILINESTRING ((-0.931034...   2 transition     28
#> 84 MULTIPOINT ((-0.9464064 -0....   3 transition     28
#> 85 MULTIPOLYGON (((0.2012327 0...   1 transition     29
#> 86 MULTILINESTRING ((-0.965517...   2 transition     29
#> 87 MULTIPOINT ((-0.9732032 -0....   3 transition     29
#> 88 MULTIPOLYGON (((0 1.25, 0.0...   1        raw     30
#> 89 MULTILINESTRING ((-1 0.3686...   2        raw     30
#> 90 MULTIPOINT ((-1 -1), (-1 -0...   3        raw     30