Some different geometries to play with

These functions are provided to allow you to play with somee simple shapes as you explore transformr and are also used in the examples for the different tween functions. All geometries can be returned as either a standard data.frame with x, y, and id column, or as an sf geometry of the appropriate type.

poly_circle(st = FALSE, detail = 360)

poly_circles(st = FALSE, n = 3, r = 0.25, detail = 360)

poly_star(st = FALSE, n = 5, r1 = 0.5)

poly_star_hole(st = FALSE, n = 5, r1 = 0.5)

path_spiral(st = FALSE, windings = 5)

path_waves(st = FALSE, w1 = 7, w2 = 11)

point_random(st = FALSE, n = 10)

point_grid(st = FALSE, dim = 5)

Arguments

st

Logical. Should the geometry be returned as an sf feature?

detail

The number of points defining the shape

n

For poly_circles the number of circles, for poly_star and poly_star_hole the number of 'arms', and for point_random the number of points

r, r1

The radius of the geometry. r gives the radius of the circles in poly_circles and r1 gives the inner radius for poly_star/poly_star_hole, thus determining how pointy it is

windings

The number of revolutions in the spiral

w1, w2

The frequency for the two sine waves

dim

the number of rows and columns in the grid

Value

Either a data.frame or an sf feature depending on the value of st

Examples

# Create a 7-pointed star
poly_star(n = 7)
#>                x          y id
#> 1   0.000000e+00  1.0000000  1
#> 2   2.169419e-01  0.4504844  1
#> 3   7.818315e-01  0.6234898  1
#> 4   4.874640e-01  0.1112605  1
#> 5   9.749279e-01 -0.2225209  1
#> 6   3.909157e-01 -0.3117449  1
#> 7   4.338837e-01 -0.9009689  1
#> 8   6.123234e-17 -0.5000000  1
#> 9  -4.338837e-01 -0.9009689  1
#> 10 -3.909157e-01 -0.3117449  1
#> 11 -9.749279e-01 -0.2225209  1
#> 12 -4.874640e-01  0.1112605  1
#> 13 -7.818315e-01  0.6234898  1
#> 14 -2.169419e-01  0.4504844  1